requirements.txt
# Core dependencies
opencv-python>=4.5.0
Pillow>=8.0.0
python-dotenv>=0.19.0
requests>=2.25.0

# AI/ML dependencies
torch>=1.9.0
torchvision>=0.10.0
git+https://github.com/openai/CLIP.git

# API clients
openai>=1.0.0
google-generativeai>=0.3.0

# Optional: For better performance
numpy>=1.21.0 